changes to QexpConfig: - move from config.py to new qexp_config.py - add to __init__.py for gamechangerml/configs - refactor MODEL_ARGS -> INIT_ARGS, EXPANSION_ARGS, BUILD_ARGS - update references to this class - add docstrings
