+++ gamechangerml/configs/__init__.py
@@ -4,3 +4,4 @@ from .embedder_config import EmbedderConfig
 from .doc_compare_embedder_config import DocCompareEmbedderConfig
 from .similarity_config import SimilarityConfig
 from .doc_compare_similarity_config import DocCompareSimilarityConfig
+from .qexp_config import QexpConfig
