+++ gamechangerml/scripts/run_evaluation.py
@@ -6,8 +6,10 @@ from gamechangerml.src.model_testing.evaluation import (
     NLIEvaluator,
     QexpEvaluator,
 )
-from gamechangerml.configs import QAConfig, EmbedderConfig, SimilarityConfig
-from gamechangerml.configs.config import (
+from gamechangerml.configs import (
+    QAConfig,
+    EmbedderConfig,
+    SimilarityConfig,
     QexpConfig,
 )
 from gamechangerml.src.utilities.test_utils import *
@@ -107,8 +109,8 @@ def eval_sim(model_name, sample_limit, eval_type="original"):
 def eval_qe(model_name):
     domainEval = QexpEvaluator(
         qe_model_dir=os.path.join(MODEL_PATH, model_name),
-        **QexpConfig.MODEL_ARGS["init"],
-        **QexpConfig.MODEL_ARGS["expansion"],
+        **QexpConfig.INIT_ARGS,
+        **QexpConfig.EXPANSION_ARGS,
     )
     results = domainEval.results
     logger.info(f"Evals: {str(results)}")
@@ -176,8 +178,8 @@ def _qexp(limit):
     logger.info("\nEvaluating Query Expansion with GC data...")
     QEEval = QexpEvaluator(
         qe_model_dir=os.path.join(MODEL_PATH, "qexp_20201217"),
-        **QexpConfig.MODEL_ARGS["init"],
-        **QexpConfig.MODEL_ARGS["expansion"],
+        **QexpConfig.INIT_ARGS,
+        **QexpConfig.EXPANSION_ARGS,
     )
     logger.info(QEEval.results)
