+++ gamechangerml/train/pipeline.py
@@ -51,12 +51,10 @@ from gamechangerml.api.utils.pathselect import get_model_paths
 
 from gamechangerml.src.search.query_expansion.build_ann_cli import build_qe_model as bqe
 from gamechangerml.src.utilities import utils
-from gamechangerml.configs import EmbedderConfig, SimilarityConfig
+from gamechangerml.configs import EmbedderConfig, SimilarityConfig, QexpConfig
 from gamechangerml.configs.config import (
     DefaultConfig,
     D2VConfig,
-    QexpConfig,
-    QexpConfig,
 )
 
 import pandas as pd
@@ -417,7 +415,7 @@ class Pipeline:
         try:
             # build ANN indices
             index_dir = os.path.join(model_dest, model_path)
-            bqe.main(corpus, index_dir, **QexpConfig.MODEL_ARGS["bqe"])
+            bqe.main(corpus, index_dir, **QexpConfig.BUILD_ARGS)
             logger.info(
                 "-------------- Model Training Complete --------------")
             # Create .tgz file
@@ -435,7 +433,7 @@ class Pipeline:
                 logger.info(
                     "-------------- Running Assessment Model Script --------------"
                 )
-                # qxpeval = QexpEvaluator(qe_model_dir=index_dir, **QexpConfig.MODEL_ARGS['init'], **QexpConfig.MODEL_ARGS['expansion'], model=None)
+                # qxpeval = QexpEvaluator(qe_model_dir=index_dir, **QexpConfig.INIT_ARGS, **QexpConfig.EXPANSION_ARGS, model=None)
                 # evals = qxpeval.results
 
                 logger.info(
